# Conflicts:
#	t1g-distro/general/ddls/local_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/t1t_acc_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/t1t_dev_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/t1t_prod_api_engine_postgres_inserts.dml
#	t1g-distro/general/ddls/update.sql

+++ t1g-ejb/src/test/java/com/t1t/apim/kong/KongClientIntegrationTest.java
@@ -361,7 +361,7 @@ public class KongClientIntegrationTest {
     @Test
     public void testGetAllPlugins() throws Exception {
         KongPluginConfigList confList = kongClient.getAllPlugins();
-        int initSize = confList.getTotal();
+        long initSize = confList.getTotal();
         print(confList);
         //add some api config
         KongApi apie = createDummyApi("apiexx", "/apiexx", API_URL);
@@ -374,7 +374,7 @@ public class KongClientIntegrationTest {
         print(pluginConfig);
         //verify one has been added
         confList = kongClient.getAllPlugins();
-        int sizeAfter = confList.getTotal();
+        long sizeAfter = confList.getTotal();
         log.info("size before:{}",initSize);
         log.info("size after:{}", sizeAfter);
         kongClient.deleteApi(apie.getId());
@@ -413,7 +413,7 @@ public class KongClientIntegrationTest {
         pluginConfig = kongClient.createPluginConfig(apif.getId(), pluginConfig);
         print(pluginConfig);
         //update
-        KongPluginRateLimiting rateConfig = new KongPluginRateLimiting().withMinute(5);
+        KongPluginRateLimiting rateConfig = new KongPluginRateLimiting().withMinute(5L);
         pluginConfig.setConfig(rateConfig);
         print(pluginConfig);
         kongClient.updateOrCreatePluginConfig(apif.getId(), pluginConfig);
@@ -691,7 +691,7 @@ public class KongClientIntegrationTest {
     private KongPluginConfig createTestPlugin(KongConsumer consumer) {
         //create config value - 1 request/minute
         KongPluginRateLimiting rateLimitingConfig = new KongPluginRateLimiting()
-                .withMinute(1);
+                .withMinute(1L);
         KongPluginConfig pluginConfig = new KongPluginConfig()
                 .withConsumerId(consumer.getId())
                 .withName("rate-limiting")//as an example
